pyproject.toml
[tool.poetry]
name = "api"
version = "1.0.1"
description = ""
authors = ["ionite34 <dev@ionite.io>"]

[tool.poetry.dependencies]
python = ">=3.10, <3.11"
google-cloud-datastore = "^2.8.1"
fastapi = "^0.79.0"
fastapi-camelcase = "^1.0.5"
uvicorn = "^0.18.2"
rolldet = ">=0.1.1"

[tool.poetry.dev-dependencies]
pytest = "^7.1.2"
pytest-cov = "^3.0.0"
black = "^22.6.0"
isort = "^5.10.1"
flake8 = "^5.0.4"
mypy = "^0.971"
pylint = "^2.14.5"
poethepoet = "^0.16.0"
requests = "^2.28.1"
pre-commit = "^2.20.0"

[tool.pylint.messages_control]
max-line-length = 88
disable = [
    "missing-module-docstring",
    "unused-argument",
    "no-value-for-parameter",
    "no-member",
    "no-else-return",
    "bad-whitespace",
    "bad-continuation",
    "line-too-long",
    "fixme",
    "protected-access",
    "too-few-public-methods",
]

[tool.pylint.design]
max-returns = 8

[tool.poe.tasks]
autoformat.sequence = [
    {cmd = "black src tests"},
    {cmd = "isort src tests"},
]

lint.sequence = [
    {cmd = "black --check -v src tests"},
    {cmd = "isort --check -v src tests"},
    {cmd = "flake8 src tests"},
    {cmd = "mypy --pretty src tests"},
]

test.sequence = ["unit", "integration"]

[tool.poe.tasks.unit]
sequence = [
    {cmd = "pytest -vv --cov-report xml --cov-report term --cov=api tests/unit"},
]

[tool.poe.tasks.integration]
sequence = [
    {shell = "pytest -vv --cov-append --cov-report xml --cov-report term --cov=api tests/integration"},
]

[tool.poe.tasks.local-server]
sequence = [
    {shell = "uvicorn api.server:app --reload --port 8011"},
]


[build-system]
requires = ["poetry-core>=1.1.0"]
build-backend = "poetry.core.masonry.api"
